[Crypto] Initial support for OpenPGP clients
The default OpenPGP API doesn't work in AndroidX, therefore needed a workaround. There's an alternative library written in Kotlin (https://github.com/android-password-store/openpgp-ktx) but it inherits the same problem: it relies on Android preference library (an API is supposed to be independent of such a library). This workaround uses a single Dialog fragment to merge the two preferences and therefore, we have at least some control on how we are going to work with it. But I have to admit, the original OpenPGP library is not what a library should look like (neither is mine as a consequence). Related issue: https://github.com/open-keychain/open-keychain/issues/2432