[Rules] Apply rules based on user handle
Non-IFW rules will be applied only for the given user handle. Previously it only applies rules for the current user. From now on, a custom user handle is set but in most of the cases, current user handle is still used. This is useful for restoring backups for a specific user handle and fixes #114.