[AppDetails] Prioritize app ops over permissions when available
Since (changeable) system permissions on Android are largely replaced by app ops internally, app ops should have the top priority (over permissions). But if app op is not assigned to a permission, it will fall back to the grant/revoke permission option. However, this doesn't remove the necessity for a separate tab for app ops.