[Backup] Add a settings item to backup apps with KeyStore
KeyStore backup is disabled by default as restoring them will not work in many cases. According to various reports, restoring backup will work on Android 5 (most KeyStore elements are not available at that time), may work on Android 6 through 8 and will not work after Android 9. Restoring such backups in a new device or ROM requires that the backup is taken while there is no screenlock is present and that the target device has no screenlock either. Restoring in an existing ROM will not work if the screenlock is changed after taking a backup.