[Interceptor] Add activity interceptor
Intent intercept (my fork: https://github.com/MuntashirAkon/intent-intercept) is an incredibly useful app in understanding activity intents of Android apps. It can also be used to snoop into activity intents similar to man-in-the-middle in cryptography and networking. Interceptor is my effort to make it even more useful and user friendly.