[Backup] Add restore support for URI grants (SAF)
However, the changes are visible only after a reboot. The users should also make sure that the URI permissions are not granted to any apps before the pending reboot, otherwise the system may override the changes. It may be possible to ask system to read persisted uri permissions but the possibility is still under investigation.

When restoring a URI grant, AM assumes that the source package exist for the given user ID. If the package doesn't exist, undefined situation may occur. This situation may also be handled properly in future.