[ipc] Run privilege service only after local server has started
Even though local server had started properly, AM failed to run the privileged service on ADB mode due to a deadlock introduced in the local server as it required the local server itself to run the privileged service. Since the service has to be run after the local server has started but inside the singleton constructor itself, a better way is to simply return the server without synchronisation once it's not null. As a result, the initial call to the singleton would continue to execute without interfering the subsequent calls.