Remove UnAPKM activity in favour of UnAPKM API
Since newer APKM files are DRM-free, it was no longer necessary to bundle the De-DRM tool with App Manager. Instead, the UnAPKM tool (https://github.com/MuntashirAkon/unapkm-android) can be used. It provides a public API which is used by App Manager to de-DRM the old APKM files if it is installed.