[backup] Replace the flag data with internal data
Data flag was required in order to backup internal, external data folders and media/OBB files, and the internal data folder was backed up by default when the flag was enabled.  This patch replaces it with “internal data” which is only responsible for backing up the internal data folders and the other two flags have no dependency over this flag.