[Blocker] Display filenames that couldn't be imported
For importing files with Watt and Blocker, there might some situations where some files couldn't be imported (mostly due to the fact that the components no longer exist or the app is not installed). In such cases, the filenames are displayed to let the user know which ones have failed.