[Refactor] Fix support for apps uninstalled without clearing data
Neither Application#FLAG_INSTALLED nor Application#FLAG_IS_DATA_ONLY works
properly when it comes to detecting whether an app is installed. This is
specially true for applications uninstalled without clearing data. In addition
to the above checks, another check is performed to ensure that the app
directory also exists.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>