[Backup] Fix relative path issue
Java File API often fails to handle the relative files properly. As a result,
some backups have the wrong relative paths. To solve the issue, a custom
method is used with a number of tests.

The issue appears to be present only in v2.6.1. However, no backward
compatibility has been provided with this patch.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>