[KeyStore] Encrypt data using Cipher itself instead of using CipherOutputStream
Android KeyStore in older Android versions did not reset IV correctly.
As a result, when close was called on CipherOutputStream, it threw an
IllegalStateException saying that "IV has already been used. Reusing IV in
encryption mode violates security best practices." which was not correct.
This patch attempts to fix the issue by avoiding the use of CipherOutputStream
altogether.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>