[Refactor] Skip checking for remote server if no-root mode is enabled.
When no-root mode is enabled and the remote server is still running due to the
past use of ADB over TCP or Wireless Debugging, App Manager throws an error
while checking for the remote server since it is still running in the
background. This behaviour is now overridden by adding an explicit check for
no-root mode i.e., App Manager will not check for the remote server if no-root
mode is explicitly enabled.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>