[FM] Add the ability to edit/modify files via external apps
It is possible to edit or modify any files, including the files only accessible
via root, using third-party applications. It uses a custom content provider
with Linux pipe in order to achieve that. Therefore, there are many well-known
limitations including the inability to reading and writing concurrently. It
also means that the reading/writing speed is slower than the direct approach.

Despite all these limitations, sufficient care has been taken to ensure that
the file descriptor supports all the supported flags (or modes), and it is up
to the editor or viewer to ensure that the file is being read or written
correctly. Some file managers, including MiXplorer, do not truncate files i.e.
they do not use the “t” mode during writing which can cause problems if the
edited file contains less text than it initially did.

It should be ideal to replace the content provider with document provider which
might be considered later considering the lack of support for such provider by
third-party applications.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>