[Refactor] Reimplementation of ProxyInputStream
Java's default FileInputStream is unusable in App Manager due to its limitations as well as complexity. Instead, the underlying file descriptor is used directly for reading a file, effectively making it the same as FileInputStream with additional support for our proxy files. However, for root and ADB mode of operations, it reads a file remotely using a remote service. Hence, previous limitations still apply.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>