[Refactor] Reimplementation of ProxyOutputStream
Java's default FileOutputStream is unusable in App Manager due to its
limitations as well as complexity. Instead, the underlying file descriptor is
used directly for writing a file, effectively making it the same as
FileOutputStream with additional support for our proxy files. However, for root
and ADB mode of operations, it writes to a file remotely using a remote
service. Hence, the previous limitations still apply.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>