[Refactor] Fix navigation theming issue in Android < 27
Android SDK versions < 27 only support changing the colours of the navigation
background, not the buttons. As a result, when the white background is set in
light mode, the navigation buttons become more or less invisible.

To circumvent the issue, the white background is replaced by the accent colour
for Android SDK versions < 27. For 27 or later, the white background is still
used.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>