[Interceptor] Enable copy/paste
Supported data from the interceptor can be copied, persisted and pasted later.
Supported data includes primitives such as integer, long integer, floating
point number, string and their arrays. Other data types are not supported and
will be dropped when the intent is being flattened to string. The generated
output contains redundant but useful data such as matching activities and
result intent (if any) which are completely optional and is dropped when pasted
from the clipboard.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>