[AppDetails] Enable opening non-exported activities via interceptor
Similar to the exported activities, non-exported activities can also be opened
via interceptor by long clicking on an activity name with root. For an activity
with a permission not granted to App Manager, the activity will be opened with
root. Currently, it's not possible to display result for the activities
launched with root since the ActivityManagerCompat class has not been modified
to handle request code.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>