[AppDetails] Fetch running services from activity manager
Previously, `dumpsys activity services` was used to retrieve the list of
running services and were matched with the running apps to fetch process IDs.
While dumpsys command could be accessed with android.permission.DUMP without
root/ADB, fetching a list of running apps required root/ADB. This is now
handled in two different ways:
1. In no-root mode, android.permission.DUMP permission (if granted) is still
   used to fetch the running services, but this time, the process IDs are
   fetched from the dump itself rather than fetching running apps which
   required a more privileged permission.
2. In root/ADB mode, the running services are fetched directly from
   `ActivityManager#getServices()`.

In the App Info tab, clicking on any running service opens the log viewer with
process ID as the filter. Other behaviours are continued to be the same.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>