[Refactor] Fetch running apps/services directly when required
In no-root mode, if android.permission.DUMP is not granted, App Manager will
use the ActivityManager to fetch running processes and services. For Android
Lollipop, this will return all running processes and services, and for Android
M and later, this will only return the processes and services under the UID
where App Manager is running (currently, this means that it will list App
Manager only).

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>