[Installer] Add an info icon to open app in app details page
After the installation is finished successfully, an info button is displayed on
the top-right of the dialog. Clicking on the dialog opens the regular app
details page for the app. However, on clicking the back button on the page, it
goes to the main page instead of exiting the app.

Additional fix includes displaying version and tracker info in the “changes”
(or “What's New”) dialog.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>