[Refactor] Cache files in the cache directory
Previously, files were cached in the data/cache directory which would increase
the data size if the cached APK files weren't deleted for some reason.

Additionally, the whole cache directory was cleared everytime the app details
page was visited which totally destroyed the purpose of using caches and
therefore, was removed.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>