[AppInfo] Use accessibility to clear data/cache and force-stop apps
In non-root mode, it is possible to clear app cache, app data, and force-stop
apps using Android accessibility service. However, App Manager does not provide
an interface to configure this option. The only way to configure accessibility
is by turning it on explicitly from Android Settings > Accessibility.

Accessibility service is tested only on Android 11. It should also work on a
few other versions, but no guarantee has been given. Additionally, this service
should only work on the AOSP or the AOSP-flavoured operating systems, other
operating systems such as MIUI, EMUI, One UI, Color OS are not supported
(and may not be supported in any time soon).

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>