[AppDetails] Enable advanced options for disabling components
On long clicking on the block/unblock button, a context menu would appear where
the user can block/enable/disable a component in various ways, such as--
1. Block using Intent Firewall (IFW) and PackageManager (PM)
2. Block using IFW only
3. Disable using PM
4. Enable using PM
5. Set to default (as specified in the manifest) via PM

It should be noted that, operation no. 2 does not change component status i.e.,
if it is previously disabled, it will remain disabled. However, all actions
modify the IFW rules (if applicable) as they should be.

Regarding tests, the tests have not been updated to include the new scheme.
As a result, there is no guarantee that the new scheme is foolproof.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>