[AppInfo] Add new tag: Play App Signing
The apps having the Play App Signing markers are tagged “Pay App Signing”.
On clicking the tag, a dialog is displayed where the tag is explained. Since
this also indicates that the signing key is uploaded to Google servers i.e.
a cloud infrastructure, the tag is coloured red to warn the user of potential
danger in case of a breach in the cloud infrastructure. (Hacking incidents in
2019-21 proved that no cloud infrastructure is secure enough.)

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>