[RunningApps] Display process details
It is possible to view the details of a process in a popup dialog by clicking
on it. As always, everything in the dialog is selectable. The new dialog
displays memory and CPU usage along with priority, states, thread count,
SELinux context, commandline arguments and so on. For Android apps, it is also
possible to launch the App Details page by clicking on the “info” icon in the
top-right corner of the dialog.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>