[MagiskHide] Support for multi-process apps
Like the Magisk app, it is possible to hide Magisk root from the selected
processes in a multi-process app in the App Info page. It is also possible to
back up and restore them (and this does not depend on whether they are
configured within App Manager or not) when “Back up extras” option is selected.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>