[API] Add the ability to trigger profiles via third-party apps
In order to ensure security, an authorization key is introduced which can be
accessed from Settings > Authorization Manager. This key can be used to trigger
profiles via any third-party apps such as Automation or Tasker.

Specification:
`io.github.muntashirakon.AppManager.crypto.auth.AuthFeatureDemultiplexer` is
the activity responsible for handling profiles (or any other features that
might be introduced in the future). Sending an Intent to the activity performs
the designated operation by redirecting the Intent to the desired activity or
service accordingly. It has two primary (and required) extras namely `auth` and
`feature`. `auth` contains the authorization key generated earlier, and
`feature` currently accepts the value `profile` for running profiles. In the
future, more features will be added.

To configure a profile, the following extras have to be set as well:
- `prof`: The name of the profile as displayed in the Profiles page. This is a
  required field for the feature.
- `state`: State of the profile -- `on` or `off` -- as specified in the
  documentation. If this extra is not set, App Manager will display a prompt
  where a state must be selected. Therefore, for complete automation, this
  option should be set.

This authorization technique, while not new, is not entirely secure. Ideally,
the third-party apps should only be able to use the functions of another app if
they declare a certain permission required by the latter app. For automation
apps, this is not possible since they do not support per-app permission.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>