[SharedPref] Add support for sets
`set` tag contains a list of `string` tags. In App Manager shared preferences
editor, this is known as the “set of string” and the strings are separated by
comma. Since comma has a special character, if it appears in any string, it
must be escaped.

Example:

  Test string,String\,2

This would be converted as follows:

  <set>
    <string>Test string</string>
    <string>String,2</string>
  </set>

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>