[Setting] Block user until ADB connection is fully configured
After enabling ADB over TCP or Wireless Debugging, due to a lack of user
interaction, the user may resolve to think that no additional interaction is
required which may or may not be correct. Instead, a progress dialog is
displayed while it processes the mode of operation in order to block the user
from interacting with App Manager.

It also removes a few deprecated methods of interacting with ADB.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>