[Docs] Support redirecting to the locale specific links if JavaScript is enabled
In App Manager website, it is possible to redirect to the user's preferred
locale by specifying relative link when JavaScript is enabled.

Example: To navigate to the documentations related to App Details page, the
following link can be used without specifying any locale:

    https://muntashirakon.github.io/AppManager/#sec:app-details-page

If the user has already set a locale in the browser (say, German), the link
will be converted to a permanent link like below and will be redirected
automatically:

    https://muntashirakon.github.io/AppManager/de/#sec:app-details-page

If no locale is set, the user can click on any locale from the list and the
permanent link (like above) will still be resolved.

However, this only works if JavaScript is enabled, and it does not matter if
cookies are enabled or not.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>