[Interceptor] Add option to copy as `am` command
`am` (short for activity manager) or `cmd activity` (Android 9 and later) can
be used to start an activity from the terminal. This option let user copy the
Intent as an `am` command so that the same Intent can be invoked in the
terminal. But not all extras are supported, only the supported extras are
copied during the operation.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>