[Refactor] Use unprivileged service to fetch package size info
In Android 7.1 or earlier, shell may not include GET_PACKAGE_SIZE in its
manifest and may crash in some situations. Since this permission is a normal
permission, the package size info is fetched using an unprivileged service
instead of a privileged service whenever it can.

In the latter versions, it is also checked if the app is granted the permission
GET_USAGE_ACCESS before fetching the size info of a package.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>