[Installer] Use a separate notification channel for confirming installation
In no-root mode, installation confirmation notification is displayed if the
app is installed in the background. This confirmation notification might not
be displayed if another app was queued for installation immediately after the
confirmation notification is displayed, and the user didn't have enough time to
click on the notification to display the confirmation dialog. As a result, the
installation process will run several hours until it finally times out. The
issue is fixed by using a separate notification channel for the confirmation
notifications, thereby, preventing any chance of missing the confirmation
notification.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>