[Refactor] Replace SLF4J dependency with Android compatible classes
Jadx has SLF4J dependency which should not be used in Android. There are a few
Android wrappers for SLF4J, but there is no point in using them since they
increase the size of the app and Android already has built-in options for
logging. Instead, essential files in SLF4J API are replaced with Android
-compatible logging alternatives.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>