Add build expiry date
To ensure the safety of the users, App Manager is set to expire several months
after it is published. A warning is issued for a certain period before the
expiry date. After that, App Manager shall completely stop responding to any
requests and prompt user to update or uninstall the app.

    Build    Timespan   Warning period
    -----    --------   --------------
    Debug    2 months   2 weeks
    Alpha    6 months   1 month
    Beta     6 months   1 month
    RC       6 months   1 month
    Stable   18 months  3 months

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>