[Usage] Attempt to fix the usage time calculation method
The usage events returned by the UsageStatsManager was assumed to be in order of
their timestamp which does not seem to be true and resulted in missing a few
events due to the standard calculation method, that is, calculating the time
difference between activity resume and pause time. In addition, it appears that
the system may log activity stop time without logging any pause time (a typical
cycle would be resume --> pause --> stop) causing further miss of events. All of
these issues are addressed by sorting the events in order of their timestamp as
well as measure the time difference between resume and stop times instead of
resume and pause times. However, this does not solve the problems with access
count since it is still is a summation of the resume-stop cycles.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>