[Usage] Attempt to fix the “times opened” value
In order to calculate a more reliable number of times an app was opened, the
time difference between each activity opening and closing are now relaxed. So,
when the user navigates to another activity from an activity belonging to the
same app and the time difference is less than 500 ms, the time difference is
added to the total usage time. This calculation is technically more precise than
the system's own open count because certain navigation involves the use of
system UI (e.g., the arrow in the gesture navigation is emitted from the System
UI app and should be considered as such) which are ignored. However, the issues
with choosing 500 ms (half of a second) time difference are as follows:
1. Due to an user waiting a long time to trigger the back gesture or simply bad
   coding, an app may take more than this time to open an activity which will
   cause them to be listed as two accesses.
2. Some apps support multiple windows. If the user opens another activity in a
   new window shortly after opening the first activity, the two accesses may be
   counted as one instead of two.
These events are considered extraordinary and can be negligible in a real-life
setting.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>