Open Audio Player in another unprivileged process
Even if App Manager runs in the privileged mode, the Audio Player will run in a
separate unprivileged process, reducing attack surface for third-party malicious
apps as well as preventing the player from closing in case the system runs out
of resources (the latter is still in progress).

Note: The process retains all the permissions granted to App Manager. So, it's
NOT an effective process isolation technique, and various attack surface are
still present. In an ideal process isolation technique, the audio should be
played using an isolated service with limited permissions. Such technique will
be explored in the future.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>