[Refactor] Use alphanumeric sort instead of alphabetical sort
Alphabetical sort does not understand numerals. Therefore, name1, name2, name10
are sorted as name1, name10, name2 which is not preferable in some cases such as
ordering the files in a directory. Instead, alphanumeric sort is used.

For alphanumeric sort, we have borrowed the implementation by David Koelle[1]
instead of the traditional, expensive regular expression-based implementation.

[1]: http://www.davekoelle.com/alphanum.html

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>