[Finder] Add support for logical expressions
Each filter option is assigned a unique ID consisting of its type and a number.
These unique IDs acts as operands or variables in the expression. The expression
supports the following operations in order:
- Brackets: Any operand or expression inside the first bracket will be evaluated
  first. Brackets can also be nested.
- And: Logical "and" operation can be performed by separating multiple operands
  by & (ampersand) symbol. This is the default behavior when no expression is
  explicitly specified.
- Or: Logical "or" operation can be performed using the | (pipe) symbol.

All of above follows the standard Boolean algebraic laws for evaluation. Not
operator is not supported, because it already has implicit support in each
filter.

Example: Consider the following expression:

  apk_size_1 & (app_type_2 | bloatware_3)

For each application item, app_type_2 filter will be matched first as it is the
first item inside the brackets. If it returns true (i.e., the filter matched),
it shall skip checking bloatware_3 filter due to short-circuit and collapse the
expression as stated below. If it returns false, it will evaluate bloatware_3
and collapse the expression.

  apk_size_1 & true

In this simplified form, first apk_size_1 will be evaluated. If it returns true,
the second operand (i.e., true) will be evaluated and the expression will return
true. If it returns false, the second operand will not be evaluated and the
expression will return false.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>