Infer actual battery capacity from current capacity
Previously, the "About the device" displayed only the current battery capacity
which is not the actual battery capacity. The actual battery capacity can be
calculated as follows:

(battery capacity x 100 ) / (battery percentage)

However, the battery percentage supplied by BatteryManager is a whole number.
Therefore, there will be a rounding off error. This error is much low when the
battery percentage is near 100%.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>