Build.VERSION_CODES.O Seems to be the level at which overlays were really added, the Auto complete claims Build.VERSION_CODES.UPSIDE_DOWN_CAKE which is complete lie.
But is now working last tested on rooted devices further testing with adb is required. RequiredApi information will have to be made
It's not Required for IOverlayManager.java as that is the first commit of that file in AOSP. but OverlayInfoHidden.java will have to be updated with required @RequiredApi for methods and fields.

### ProxyBinder.java Remote Fix
Was Updated to Support Loading Overlay Service and others from IAMService.aidl instead of from AM context as special permissions are required to be held for ServiceManager.java to admit that the service exists the overlay service is gated by SELinux it's allowed contexts are id 0, android.uid.shell, android.uid.system, android.uid.systemui, via assoc permission android.permission.CHANGE_OVERLAY_PACKAGES, the comment in the docs about only needing android.permission.INTERACT_ACROSS_USERS(or _FULL) seems to be a complete lie but further testing should be done as SELinux isn't nice enough to tell us why it denied something.