[Profile] Implement filter-based profile
Filter-based profiles are very similar to the apps-based profile except when the
profile is being applied, the list of applications are retrieved and filtered
using the filter options. To facilitate the user, a real-time filtered list of
applications are displayed in the profile (editor) page.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>
(cherry picked from commit 3d502a9527606c748e0877cb61523e564ecae816)