[AppInfo] Open "Open by default" settings in the "Open links" dialog
If App Manager does not have enough permission to manage domains or links, the
"Open links" dialog has a button that directs the user to Android Settings. If
the app has the "Open by default" page, clicking on the button will open that
page instead of the "App info" page (which is still used as a fallback method).

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>
(cherry picked from commit 279aad0b918fb9fb42bef8dbdebda4763e4b9052)