[Main] Replace most of the filter implementation with Finder-based filters
The underlying filtering mechanism used in the "List options" is altered to use
the Finder-based filtering instead of the regular cache-based filtering. This is
beneficial for many reasons, such as it allows semi-realtime update compared to
mostly cache-based update.

Currently, there are two exceptions due to non-existent alternative in Finder:
1. Storage access framework (SAF)
2. APK with splits.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>