[Refactor] Add support for advanced copy and pasting
If the size of the text being copied exceeds 1 MB, it is stored in a temporary
file and the file URI is copied instead of the text itself. During pasting, the
supplied text/URI or any other type of data is coerced to text value.

This required some nasty workarounds for Sora Editor which does not inherently
support such copy and pasting.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>