[Refactor] Improve the performance of tracker and library scanner
To improve performance of the scanner, the native (C++) implementation of the
Aho-Corasick algorithm has been adopted. Together with parallel processing,
it has resulted in up to 25x and 1000x speedup in tracker and library scanning
respectively. However, in the older versions of Android, the parallel processing
method used here may not be available which may reduce performance by at least
1.5x depending on the available CPU cores and threads.

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>