backup: add support for ADB backup
A new backup option, "ADB data", has been added. This option supersedes the
"Internal data" option and the "External data" option. If the option is enabled,
an AB formatted backup is taken instead of internal or external storage backup.
The backup file takes the usual data<numeric> filename with .ab extension (and
other extensions depending on the encryption method).

Signed-off-by: Muntashir Al-Islam <muntashirakon@riseup.net>